Fix compilation on 32bit platforms

+++ include/inpafile.hpp
@@ -5,7 +5,7 @@
 #include <map>
 
 // GStreamer requires g_malloc instead of ::operator new because it deallocates memory with g_free
-void* DefaultAlloc(uint64_t Size);
+void* DefaultAlloc(size_t Size);
 
 class INpaFile : NpaFile
 {
@@ -23,13 +23,13 @@ public:
 
     char* ReadFile(const std::string& Filename, uint32_t& Size);
     char* ReadFile(NpaIterator iter);
-    char* ReadFile(NpaIterator iter, uint32_t Offset, uint32_t Size, void *(*Alloc)(uint64_t) = DefaultAlloc);
+    char* ReadFile(NpaIterator iter, uint32_t Offset, uint32_t Size, void *(*Alloc)(size_t) = DefaultAlloc);
     NpaIterator Begin() { return Registry.begin(); }
     NpaIterator End() { return Registry.end(); }
     NpaIterator FindFile(const std::string& Filename) { return Registry.find(Filename); }
 private:
     void ReadHeader();
-    char* ReadData(uint32_t GlobalOffset, uint32_t LocalOffset, uint32_t Size, void *(*Alloc)(uint64_t));
+    char* ReadData(uint32_t GlobalOffset, uint32_t LocalOffset, uint32_t Size, void *(*Alloc)(size_t));
 
     std::map<std::string, Entry> Registry;
 };
